+++ lib/typescript/index.d.ts
@@ -1,0 +1,7 @@
+export { default as CollapsibleTabView } from './CollapsibleTabView';
+export type { Props as CollapsibleTabViewProps, RenderTabBarProps, } from './CollapsibleTabView';
+export { default as useCollapsibleScene } from './useCollapsibleScene';
+export type { CollapsibleScenePropsAndRef } from './types';
+export { createContext } from './CollapsibleTabViewContext';
+export { default as MaterialTopTabsCollapsibleTabView } from './MaterialTopTabsCollapsibleTabView';
+export { default as createMaterialCollapsibleTopTabNavigator } from './createMaterialCollapsibleTopTabNavigator';